% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{cb_palettes}
\alias{cb_palettes}
\title{List of all Colour-Blind Friendly palettes}
\format{
An object of class \code{list} of length 13.
}
\usage{
cb_palettes
}
\description{
List of all Colour-Blind Friendly palettes
}
\examples{
library(unikn)
unikn::seecol(cb_palettes, main = "All CB-friendly Palettes", pal_names = names(cb_palettes))
}
\keyword{datasets}
