% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{div_BuRd}
\alias{div_BuRd}
\title{Colour-Blind Friendly Diverging Blue-Red Palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
div_BuRd
}
\description{
Colour-Blind Friendly Diverging Blue-Red Palette
}
\examples{
unikn::seecol(div_BuRd, main = "div_BuRd")
}
\keyword{datasets}
