% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{div_nightfall}
\alias{div_nightfall}
\title{Colour-Blind Friendly Diverging Nightfall Palette}
\format{
An object of class \code{character} of length 17.
}
\usage{
div_nightfall
}
\description{
Colour-Blind Friendly Diverging Nightfall Palette
}
\examples{
unikn::seecol(div_nightfall, main = "div_nightfall")
}
\keyword{datasets}
