% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{div_PrGn}
\alias{div_PrGn}
\title{Colour-Blind Friendly Diverging Purple-Green Palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
div_PrGn
}
\description{
Colour-Blind Friendly Diverging Purple-Green Palette
}
\examples{
unikn::seecol(div_PrGn, main = "div_PrGn")
}
\keyword{datasets}
