% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{div_sunset}
\alias{div_sunset}
\title{Colour-Blind Friendly Diverging Sunset Palette}
\format{
An object of class \code{character} of length 11.
}
\usage{
div_sunset
}
\description{
Colour-Blind Friendly Diverging Sunset Palette
}
\examples{
unikn::seecol(div_sunset, main = "div_sunset")
}
\keyword{datasets}
