% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Azure_Results.R
\name{getAzureResults}
\alias{getAzureResults}
\title{Connect to the Azure Database and Run a Query}
\usage{
getAzureResults(SQLscript, output, overwrite = FALSE)
}
\arguments{
\item{SQLscript}{a glue object containing a SQL script (from glue_sql())}

\item{output}{a string containing the path for a csv output}

\item{overwrite}{a logical indicating whether or not the output csv should be overwritten. Defaults to FALSE.}
}
\value{
a tibble containing the results of the SQL query.
}
\description{
If overwrite = FALSE and the output file already exists, the database will not be queried but the csv read instead.
}
