% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{likert3}
\alias{likert3}
\title{Unifrog likert3 palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
likert3
}
\description{
Unifrog likert3 palette
}
\examples{
library(unikn)
seecol(likert3, main = "likert3")
}
\keyword{datasets}
