% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{likert5}
\alias{likert5}
\title{Unifrog likert5 palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
likert5
}
\description{
Unifrog likert5 palette
}
\examples{
unikn::seecol(likert5, main = "likert5")
}
\keyword{datasets}
