% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{likert7}
\alias{likert7}
\title{Unifrog likert7 palette}
\format{
An object of class \code{character} of length 7.
}
\usage{
likert7
}
\description{
Unifrog likert7 palette
}
\keyword{datasets}
