% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\name{palette_unifrog}
\alias{palette_unifrog}
\title{Unifrog colour palette}
\usage{
palette_unifrog(palette = "unifrog_main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette. Depending on the color scale, can
be `"main"` (default, discrete), or `"likert3"`, `"likert5"`, `"likert7"` for diverging scales.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to [`colorRampPalette()`][colorRampPalette].}
}
\description{
The palette is based on the colours for Unifrog tools as shown in the tool.tool_hex on Azure.
}
\details{
This function is usually not called directly, but from within
  [`scale_color_unifrog()`][scale_color_unifrog].
}
