% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{qual_bright}
\alias{qual_bright}
\title{Colour-Blind Friendly Qualitative Bright Palette}
\format{
An object of class \code{character} of length 7.
}
\usage{
qual_bright
}
\description{
Colour-Blind Friendly Qualitative Bright Palette
}
\examples{
unikn::seecol(qual_bright, main = "qual_bright")
}
\keyword{datasets}
