% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{qual_hc}
\alias{qual_hc}
\title{Colour-Blind Friendly Qualitative High-Contrast Palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
qual_hc
}
\description{
Colour-Blind Friendly Qualitative High-Contrast Palette
}
\examples{
unikn::seecol(qual_hc, main = "qual_hc")
}
\keyword{datasets}
