% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{qual_light}
\alias{qual_light}
\title{Colour-Blind Friendly Qualitative Light Palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
qual_light
}
\description{
Colour-Blind Friendly Qualitative Light Palette
}
\examples{
unikn::seecol(qual_light, main = "qual_light")
}
\keyword{datasets}
