% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{qual_mc}
\alias{qual_mc}
\title{Colour-Blind Friendly Qualitative Medium Contrast Palette}
\format{
An object of class \code{character} of length 6.
}
\usage{
qual_mc
}
\description{
Colour-Blind Friendly Qualitative Medium Contrast Palette
}
\examples{
unikn::seecol(qual_mc, main = "qual_mc")
}
\keyword{datasets}
