% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{qual_muted}
\alias{qual_muted}
\title{Colour-Blind Friendly Qualitative Muted Palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
qual_muted
}
\description{
Colour-Blind Friendly Qualitative Muted Palette
}
\examples{
unikn::seecol(qual_muted, main = "qual_muted")
}
\keyword{datasets}
