% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{qual_pale}
\alias{qual_pale}
\title{Colour-Blind Friendly Qualitative Pale Palette}
\format{
An object of class \code{character} of length 6.
}
\usage{
qual_pale
}
\description{
Colour-Blind Friendly Qualitative Pale Palette
}
\examples{
unikn::seecol(qual_pale, main = "qual_pale")
}
\keyword{datasets}
