% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{qual_vibrant}
\alias{qual_vibrant}
\title{Colour-Blind Friendly Qualitative Vibrant Palette}
\format{
An object of class \code{character} of length 7.
}
\usage{
qual_vibrant
}
\description{
Colour-Blind Friendly Qualitative Vibrant Palette
}
\examples{
unikn::seecol(qual_vibrant, main = "qual_vibrant")
}
\keyword{datasets}
