% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\name{scale_color_unifrog}
\alias{scale_color_unifrog}
\alias{scale_color_unifrog_d}
\alias{scale_color_unifrog_c}
\alias{scale_colour_unifrog}
\alias{scale_colour_unifrog_c}
\alias{scale_colour_unifrog_d}
\alias{scale_fill_unifrog}
\alias{scale_fill_unifrog_d}
\alias{scale_fill_unifrog_c}
\title{Unifrog colour palette}
\usage{
scale_color_unifrog(
  palette = "qual_vibrant",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_unifrog_d(
  palette = "qual_vibrant",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_unifrog_c(
  palette = "seq_iridescent",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_unifrog(
  palette = "qual_vibrant",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_unifrog_c(
  palette = "seq_iridescent",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_unifrog_d(
  palette = "qual_vibrant",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_fill_unifrog(
  palette = "qual_vibrant",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_unifrog_d(
  palette = "qual_vibrant",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_unifrog_c(
  palette = "seq_iridescent",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)
}
\arguments{
\item{palette}{Character name of palette. Depending on the color scale, can
be `"main"` (default, discrete), or `"likert3"`, `"likert5"`, `"likert7"` for diverging scales.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{aesthetics}{A vector of names of the aesthetics that this scale
should be applied to (e.g., `c('color', 'fill')`).}

\item{...}{Additional arguments passed to `discrete_scale()` when `discrete`
is `TRUE` or to `scale_color_gradientn()` when `discrete` is `FALSE`.}
}
\description{
The palette is based on the colours for Unifrog tools as shown in the tool.tool_hex on Azure.
Use `scale_color_unifrog_d` for *discrete* categories and
`scale_color_unifrog_c` for a *continuous* scale.
}
\examples{
library(ggplot2)

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_boxplot() +
  scale_fill_unifrog_d() +
  theme_bw()

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_violin() +
  scale_fill_unifrog_d(palette = "likert3") +
  theme_bw()

ggplot(iris, aes(x = Petal.Length, y = Petal.Width, color = Sepal.Length)) +
  geom_point() +
  scale_color_unifrog_c(palette = "likert7") +
  theme_bw()
}
