% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{seq_incandescent}
\alias{seq_incandescent}
\title{Colour-Blind Friendly Sequential Incandescent Palette}
\format{
An object of class \code{character} of length 11.
}
\usage{
seq_incandescent
}
\description{
Colour-Blind Friendly Sequential Incandescent Palette
}
\examples{
unikn::seecol(seq_incandescent, main = "seq_incandescent")
}
\keyword{datasets}
