% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{seq_iridescent}
\alias{seq_iridescent}
\title{Colour-Blind Friendly Sequential Iridescent Palette}
\format{
An object of class \code{character} of length 23.
}
\usage{
seq_iridescent
}
\description{
Colour-Blind Friendly Sequential Iridescent Palette
}
\examples{
unikn::seecol(seq_iridescent, main = "seq_iridescent")
}
\keyword{datasets}
