% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{seq_YlBr}
\alias{seq_YlBr}
\title{Colour-Blind Friendly Sequential Yellow-Brown Palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
seq_YlBr
}
\description{
Colour-Blind Friendly Sequential Yellow-Brown Palette
}
\examples{
unikn::seecol(seq_YlBr, main = "seq_YlBr")
}
\keyword{datasets}
