% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrog_theme.R
\name{set_unifrog_geoms}
\alias{set_unifrog_geoms}
\title{Set geom defaults for unifrog theme}
\usage{
set_unifrog_geoms(point = TRUE, line = TRUE, bars = TRUE, text = TRUE)
}
\arguments{
\item{point}{boolean set defaults for "point" geom}

\item{line}{boolean set defaults for "line" geom}

\item{bars}{boolean set defaults for "bar" and "col" geoms}

\item{text}{boolean set defaults for "text" and "label" geoms}
}
\description{
Set geom defaults for unifrog theme
}
