% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Azure_Results.R
\name{setupAzureCreds}
\alias{setupAzureCreds}
\title{Initial setup of Azure Keys}
\usage{
setupAzureCreds()
}
\description{
Run this after keyring is installed to set the initial values for the Azure server.
}
