% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_blues}
\alias{unifrog_blues}
\title{Unifrog blues palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
unifrog_blues
}
\description{
Unifrog blues palette
}
\keyword{datasets}
