% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\name{unifrog_colors}
\alias{unifrog_colors}
\title{Extract Unifrog colours as hex codes}
\usage{
unifrog_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the Unifrog color
palette. Use `unifrog_colors()` to see all available colors.
}
\examples{
unifrog_colors()

unifrog_colors("green", "lightblue", "darkgrey")
}
