% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_greens}
\alias{unifrog_greens}
\title{Unifrog greens palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
unifrog_greens
}
\description{
Unifrog greens palette
}
\keyword{datasets}
