% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_main}
\alias{unifrog_main}
\title{Unifrog Main palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
unifrog_main
}
\description{
Unifrog Main palette
}
\examples{
unikn::seecol(unifrog_main, main = "unifrog_main")
}
\keyword{datasets}
