% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_oranges}
\alias{unifrog_oranges}
\title{Unifrog oranges palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
unifrog_oranges
}
\description{
Unifrog oranges palette
}
\keyword{datasets}
