% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_palettes}
\alias{unifrog_palettes}
\title{List of all Unifrog palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
unifrog_palettes
}
\description{
List of all Unifrog palettes
}
\keyword{datasets}
