% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_purples}
\alias{unifrog_purples}
\title{Unifrog purples palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
unifrog_purples
}
\description{
Unifrog purples palette
}
\examples{
library(unikn)
seecol(unifrog_purples, main = "unifrog_purples")
}
\keyword{datasets}
