% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_reds}
\alias{unifrog_reds}
\title{Unifrog red palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
unifrog_reds
}
\description{
Unifrog red palette
}
\keyword{datasets}
