% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_unifrog.R
\docType{data}
\name{unifrog_teals}
\alias{unifrog_teals}
\title{Unifrog teals palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
unifrog_teals
}
\description{
Unifrog teals palette
}
\examples{
library(unikn)
seecol(unifrog_teals, main = "unifrog_teals")
}
\keyword{datasets}
