% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrog_theme.R
\name{unifrog_theme}
\alias{unifrog_theme}
\title{A ggplot theme for Unifrog employees based on theme_classic()}
\usage{
unifrog_theme(base_size = 10)
}
\arguments{
\item{base_size}{The default font_size in pts}
}
\description{
A ggplot theme for Unifrog employees based on theme_classic()
}
